refactors importer/Export of planbuilder adds yaml code to modelutils and docker container
