--- org.opentosca.planbuilder/org.opentosca.planbuilder.integration/src/main/java/org/opentosca/planbuilder/integration/layer/AbstractExporter.java
@@ -1,14 +1,0 @@
-package org.opentosca.planbuilder.integration.layer;
-
-/**
- * <p>
- * This abstract class is used to define exporters
- * </p>
- * Copyright 2013-2022 IAAS University of Stuttgart <br>
- * <br>
- *
- * @author Kalman Kepes - kepeskn@studi.informatik.uni-stuttgart.de
- */
-public abstract class AbstractExporter {
-
-}
